// src/app/core/models/class.model.ts
export interface InstructorList {
  user_id: InstructorDetail[];
  instructor_id: String;
}
export interface InstructorDetail {
  name: String;
  last_name: String;
}

export interface LabDetail {
  name: string;
  _id: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
  id: string;
}

export interface LabListModel {
  docs: LabDetail[];
}

export interface CourseTitleModel extends Pagination {
  _id: string;
  id: string;
  title: string;
}

export interface CourseListModel extends Pagination {
  docs: CourseTitleModel[];
}

export interface ClassListingModel extends Pagination {
  data: any;
  docs: ClassModel[];
  totalDocs: number;
  limit: number;
  sortBy: string;
  sortByDirection: string;
}
export interface SessionClassModel {
  _id: string;
  sessionNumber: string;
  sessionStartDate: string;
  sessionEndDate: string;
  sessionStartTime: string;
  sessionEndTime: string;
  instructorId: string;
  students: StudentRegisteredClass[];
}
export interface ClassModel {
  status: String;
  _id: String;
  courseId: any;
  courseName: String;
  classAccessType: String;
  classDeliveryType: String;
  instructorCost: String;
  instructorCostCurrency: String;
  isGuaranteedToRun : boolean;
  externalRoom:boolean;
  minimumEnrollment: String;
  maximumEnrollment: String;
  classStartDate: String;
  classEndDate: String;
  sessions: SessionClassModel[];
}
// ...
export interface Session {
  _id: string;
  instructorId: string;
  laboratoryId: string;
  sessionEndDate: string;
  sessionEndTime: string;
  sessionNumber: number;
  sessionStartDate: string;
  sessionStartTime: string;
}
// ...
export interface StudentApproval {
  studentId: string;
  classId: String;
  status: string;
  approvedOn: string;
  approvedBy: string;
  session: Session[];
}